Made changes to file layout
