+++ src/components/missions.css
@@ -58,7 +58,6 @@ thead {
 }
 
 .active {
-  background-color: #0290ff;
   color: #fff;
   padding: 0.2rem;
   border-radius: 4px;
